pyproject.toml
[project]
name = "bot-boca-socios"
version = "1.0.0"
description = "Bot para compra de entradas populares/generales en bocasocios.bocajuniors.com.ar"
requires-python = ">=3.11"
dependencies = [
    "undetected-chromedriver>=3.5.5",
    "selenium-wire>=5.1.0",
    "selenium>=4.15.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "ruff>=0.1.0",
    "pyright>=1.1.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
